Voice/requirements.txt
flask==3.1.2
flask-cors==6.0.1
openai-whisper==20250625
pydub==0.25.1
torch==2.8.0
torchaudio==2.8.0
werkzeug==3.1.3
numpy==2.2.6
requests==2.32.5
tqdm==4.67.1
